Implemented Block RAM
File "ram_sim.vhdl" was replaced by "ram_block.vhdl" and now contains
the code for the block ram. The memory controller on top of the
ram_block module is useless right now, but won't be when adding more
than one ram block to it. The assembler was updated to print out the
machine code for easy copy & paste into the block ram module.