Implemented UART TX
The output of the LEDs is currently being used as the buffer
for the UART transmission. So the LED output is send over the
UART. In the current example the name and version number is being
transmitted.

The receiver(serial_communication.py) does not print out duplicate
transmissions right now.