Updated VGA controller to read data from PROM
The VGA controller now outputs the 32x16 pixel black & white sprite
which by default represents the three letters "EPU". The sprite can
be moved to the left or to the top by pressing SW1 or SW4. Pushing both
will put the sprite in the top let corner of the screen.