Implemented VGA image display test
To display an image on the VGA screen, you have to convert the image
with the help of the coetool provided in the easm folder. The result
should be a .coe file which has to be moved to vhdl/mem/ram/init.coe to
ensure that the memory block uses the image for the VRAM. After
regenerating the block memory, the project is ready to be build and
flashed onto the FPGA.

The quality of the image will probably be worse on the screen, because
the VGA screen only supports 8-bit colors.