Fix bug that hangs input cycle
Only report a new input byte if the byte was successfully transmitted
(no frame error). For debugging purposes show the frame error indicator
on an LED.

Also make sure there is a slight delay when using the test.py script
when sending data, to prevent sending messages faster than the device is
able to handle.